ADDING SECADM SERVICE ENBLING & RULES
Adding the secadm.rules for the installed software (ex. zap).

Removing apache jails rules copy because it is now installe by vulture-gui.
This goes with removing secadm service starting that is now useless in the
install-kernel.sh.

Because of the reasons given above we do need to enable secadm earlier in the
bootstraping process.
This is why we modified some mkjails to enable and start secadm.